Move CardList to its own directory.
