+++ src/client/components/App.ts
@@ -8,7 +8,7 @@ import SpectatorHome from '@/client/components/SpectatorHome.vue';
 import {ViewModel, PlayerViewModel} from '@/common/models/PlayerModel';
 import StartScreen from '@/client/components/StartScreen.vue';
 import LoadGameForm from '@/client/components/LoadGameForm.vue';
-import CardList from '@/client/components/CardList.vue';
+import CardList from '@/client/components/cardlist/CardList.vue';
 import {SimpleGameModel} from '@/common/models/SimpleGameModel';
 import Help from '@/client/components/help/Help.vue';
 import AdminHome from '@/client/components/admin/AdminHome.vue';
